requirements.txt
alembic==1.13.3
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.8.0
blinker==1.8.2
cachetools==5.5.0
camelot-py==0.9.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
config==0.5.1
cryptography==43.0.1
et-xmlfile==1.1.0
fastapi==0.115.6
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Migrate==4.0.7
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
ghostscript==0.7
google-api-core==2.20.0
google-auth==2.35.0
google-cloud-vision==3.7.4
googleapis-common-protos==1.65.0
grpcio==1.66.2
grpcio-status==1.66.2
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
numpy==2.1.1
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.28.2
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
PyMuPDF==1.24.11
PyPDF2==2.10.5
pypdfium2==4.30.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
rsa==4.9
setuptools==75.2.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.41.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.34.0
Werkzeug==3.0.4

openai~=1.59.7
pip~=24.2
attrs~=24.3.0
distro~=1.9.0
httpcore~=1.0.7
httpx~=0.28.1
jiter~=0.8.2
aiohttp~=3.11.11
python-dotenv~=1.0.1